Rename: viewmodel, navigation 폴더 생성
